ClaimPathPointer move to utils
