+++ oid4vc-core/Cargo.toml
@@ -15,6 +15,7 @@ getset = "0.1.2"
 is_empty = "0.2.0"
 jsonwebtoken.workspace = true
 lazy_static = "1.4.0"
+nutype.workspace = true
 rand = "0.8"
 serde.workspace = true
 serde_json = "1.0"
