+++ oid4vc-core/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod authentication;
 pub mod authorization_request;
 pub mod authorization_response;
+pub mod claim_path_pointer;
 pub mod client_metadata;
 pub mod jwt;
 pub mod openid4vc_extension;
